createSendSpan -> models shared with event-hubs
